% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_fails.R
\name{co_fails}
\alias{co_fails}
\title{co_fails}
\usage{
co_fails(edges, size = 1, label_size = 1, degree_mode = "total",
  min_weight = 0, min_degree = 0)
}
\arguments{
\item{edges}{edges data.frame}

\item{size}{node sizes}

\item{label_size}{text size}

\item{degree_mode}{which degree calculates (in, out, total)}

\item{min_weight}{minimum edge weight by which to filter}

\item{min_degree}{minimum node degree by which to filter}
}
\description{
co_fails
}
