% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails_dist.R
\name{fails_dist}
\alias{fails_dist}
\title{fails_dist}
\usage{
fails_dist(edges, by = "questions")
}
\arguments{
\item{edges}{edges data.frame}

\item{by}{the subject of the distribution (questions or students)}
}
\description{
fails_dist
}
