% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_edges.R
\name{obtain_edges}
\alias{obtain_edges}
\title{obtain_edges}
\usage{
obtain_edges(answers, c_answers, avoid_a = NA, avoid_s = NA)
}
\arguments{
\item{answers}{response data.frame}

\item{c_answers}{response template}

\item{avoid_a}{questions to avoid}

\item{avoid_s}{students to avoid}
}
\description{
obtain_edges
}
