% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.otu.tab.R
\docType{data}
\name{sim.otu.tab}
\alias{sim.otu.tab}
\title{OTU count table of the simulated microbiome samples}
\format{
A data frame with 100 observations on 100 variables.
}
\usage{
data("sim.otu.tab")
}
\description{
This table contains read count data simulated for 100 samples and 100 OTUs and under the AR1 dependence structure
}
\examples{
data(sim.otu.tab)
}
\keyword{datasets}
