% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_fails.R
\name{co_fails}
\alias{co_fails}
\title{co_fails}
\usage{
co_fails(edges, size = 1, degree_mode = "total")
}
\arguments{
\item{edges}{edges data.frame}

\item{size}{node sizes}

\item{degree_mode}{which degree calculates (in, out, total)}
}
\description{
co_fails
}
