% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_edges.R
\name{obtain_edges}
\alias{obtain_edges}
\title{obtain_edges}
\usage{
obtain_edges(answers, c_answers, avoid = NA)
}
\arguments{
\item{answers}{response data.frame}

\item{c_answers}{response template}

\item{avoid}{questions to avoid}
}
\description{
obtain_edges
}
