% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_mode.R
\name{two_mode}
\alias{two_mode}
\title{two_mode}
\usage{
two_mode(edges, size = 1, label_size = 1, degree_mode = "all",
  min_degree = 0)
}
\arguments{
\item{edges}{edges data.frame}

\item{size}{node sizes}

\item{label_size}{text size}

\item{degree_mode}{which degree calculates (in, out, total)}

\item{min_degree}{minimum node degree by which to filter}
}
\description{
two_mode
}
